moved all the util classes into a util package so it doesn't clutter up home directory. Also got basic config yaml file read to work on admin
