+++ admin/config.yml
@@ -1,0 +1,11 @@
+credentials:
+  marianne:
+      clientId: abc
+      clientSecret: jQqXJmn52wupDzqSheckKu62NqpvnFuR
+      tokenURL: https://bitbucket.org/site/oauth2/access_token
+      acctName: mariannefeng
+  l11:
+      clientId: DEF
+      clientSecret: xGYhTtrDJx46kghMHujQhRSZBhHJZtwt
+      tokenURL: https://bitbucket.org/site/oauth2/access_token
+      acctName: level11consulting
