+++ admin/models/models.go
@@ -2,13 +2,18 @@ package models
 
 
 //TODO: get callback url from consul or something like it
-const WebhookCallbackURL string = "https://radiant-mesa-23210.herokuapp.com//test"
-const BuildFileName string = "README.md"
+const WebhookCallbackURL = "https://radiant-mesa-23210.herokuapp.com/test"
+const BuildFileName = "ocelot.yml"
+const ConfigFileName = "config.yml"
 
 type AdminConfig struct {
 	ConfigId     string
-	ClientId     string `validate:"required"`
-	ClientSecret string `validate:"required"`
-	TokenURL     string `validate:"required"`
-	AcctName     string `validate:"required"`
+	ClientId     string `yaml:"clientId" validate:"required"`
+	ClientSecret string `yaml:"clientSecret" validate:"required"`
+	TokenURL     string `yaml:"tokenURL" validate:"required"`
+	AcctName     string `yaml:"acctName" validate:"required"`
+}
+
+type ConfigYaml struct {
+	Credentials map[string]AdminConfig	`yaml:"credentials"`
 }
