+++ hookhandler/hookhandler.go
@@ -8,8 +8,8 @@ import (
 	"github.com/meatballhat/negroni-logrus"
 	"github.com/shankj3/ocelot/admin/handler"
 	"github.com/shankj3/ocelot/admin/models"
-	"github.com/shankj3/ocelot/nsqpb"
-	"github.com/shankj3/ocelot/ocelog"
+	"github.com/shankj3/ocelot/util/nsqpb"
+	"github.com/shankj3/ocelot/util/ocelog"
 	pb "github.com/shankj3/ocelot/protos/out"
 	log "github.com/sirupsen/logrus"
 	"github.com/urfave/negroni"
