+++ util/nsqpb/consume.go
@@ -3,7 +3,7 @@ package nsqpb
 import (
     "fmt"
     "github.com/nsqio/go-nsq"
-    "github.com/shankj3/ocelot/ocelog"
+    "github.com/shankj3/ocelot/util/ocelog"
     "os"
     "sync"
 )
